# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple service that will poll/return delegations on tezos protocol",
        "title": "Tezos Delegation Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "o.roux2@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/xtz/delegations": {
            "get": {
                "description": "Retrieve a list of delegations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get delegations",
                "operationId": "get-delegations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit the number of results (default is 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year (optional)",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.delegationJs"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.delegationJs": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "block": {
                    "type": "string"
                },
                "delegator": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "TezosAPI",
        "url": "https://api.tzkt.io/#operation/Operations_GetDelegations"
    }
}